drop support for legacy (v0) projects
This is a first in the series of PRs for cleaning up
the technical debt.
