+++ .travis.yml
@@ -22,8 +22,6 @@ install:
   -
 
 script:
-- TRACE=1 ./testv0.sh
-- ./test_existing_projects.sh
 - TRACE=1 ./testv1.sh
 
 # TBD. Suppressing for now.
